feat-137 : 최근 검색어 추가 로직 위치 SearchingFragment -> SearchingResultFragment로 수정
